requirements.txt
nltk==3.9.1
sqlparse==0.4.4
requests==2.31.0
urllib3==2.0.4
tqdm==4.66.1
parse==1.19.0
requests
python-dotenv
urllib3
